Create Voting Contract
This contract handles the voting process and locks the votes and bets for the duration of the voting period to prevent any changes. Members can only place bets if they have voted on the proposal. The contract should also track the number of votes and bets placed by each member.

Signed-off-by: naza.a.kpe.ray.mond <naza_raymond@gataprotocol.org>